refactor(docx): extract frame/float-table/anchor geometry into modules
Behavior-preserving move of self-contained geometry helpers out of the
~7.2k-line renderer.ts into three focused modules. Logic, constants, and
control flow are byte-identical; only placement changes. The existing
table-driven unit tests are the guard (import source re-pointed only).

New modules (packages/docx/src):
- frame-geometry.ts (§17.3.1.11 framePr): FrameBox, frameXContainer,
  frameYContainer, resolveAlignedPosH, resolveAlignedPosV, computeFrameBox,
  registerFrameFloat. Also hosts the shared pushFloatRect / PushFloatOpts
  exclusion-rect builder (used by frame + float-table + the renderer's
  image/shape float paths) as the base module.
- float-table-geometry.ts (§17.4.57 tblpPr / §17.4.56 tblOverlap):
  FloatTableBox, computeFloatTableBox, registerTableFloat,
  floatTableWrapSide. Reuses frame-geometry's container/align helpers and
  pushFloatRect (table↔frame anchor semantics line up 1:1).
- anchor-geometry.ts (§20.4.3.x DrawingML anchors): xContainer, yContainer,
  resolveAnchorX, resolveAnchorY.

renderer.ts imports the moved fns; resolveFrameBox / frameAnchorLineHeightPx
stay (deeply coupled to layout internals — buildSegments / layoutLines /
paragraphMarkLineHeight / DocGridCtx). RenderState is now exported as a type
so the new modules can type their params; the modules use `import type`, so no
runtime import cycle.

NICE (behavior-neutral): named MIN_ROW_HEIGHT_PT (=10pt) for the empty-row
floor in calculateRowHeight; hoisted the shared run-box rect
(effSizePx*0.85 / *1.1) into boxTop/boxHeight so highlight / §17.3.2.32
shading / §17.3.2.4 border stay byte-identical (no #541 border-merge change);
doc-noted floatTableWrapSide's column-centre side selection as a simplified
approximation.

S-11: anchor-align.test.ts inside/outside cases pin the odd-page
approximation — added a note that the true §20.4.3.1 even-page parity is
unimplemented and must update these on removal (asserts unchanged).

renderer.ts 7214 → 6627 lines. docx vitest 195/195; root typecheck green.
