+++ packages/docx/src/measure-column-geometry.test.ts
@@ -1,12 +1,12 @@
 import { describe, it, expect } from 'vitest';
+import { computeFrameBox, registerFrameFloat } from './frame-geometry.js';
 import {
-  computeFrameBox,
-  registerFrameFloat,
   computeFloatTableBox,
   registerTableFloat,
   floatTableWrapSide,
-} from './renderer.js';
-import type { FrameBox, FloatTableBox } from './renderer.js';
+} from './float-table-geometry.js';
+import type { FrameBox } from './frame-geometry.js';
+import type { FloatTableBox } from './float-table-geometry.js';
 import type { FramePr, TblpPr } from './types.js';
 import type { FloatRect } from './float-layout.js';
